Changements mineurs sur la liste

+++ app/src/main/java/ovh/tomus/iut/flotte/Models/Port.kt
@@ -2,6 +2,6 @@ package ovh.tomus.iut.flotte.Models
 
 import java.io.Serializable
 
-class Port (var id:Int, var name:String, var latitude:Float, var longitude:Float) : Serializable {
+class Port (var id:Int, var name:String, var latitude: Double, var longitude: Double) : Serializable {
 
 }
